requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.5
daphne==3.0.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
django-jalali==4.1.2
djangorestframework==3.12.2
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
jdatetime==3.6.2
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
whitenoise==5.2.0
zope.interface==5.2.0
